Convert validation tests to jest
