+++ package.json
@@ -215,7 +215,7 @@
     "github-api": "^3.0.0",
     "hast-util-sanitize": "^1.3.0",
     "highlight.js": "^9.15.6",
-    "html-inspector": "^0.8.2",
+    "html-inspector": "0.7.3",
     "htmllint": "^0.7.3",
     "i18next": "^15.0.6",
     "immutable": "^3.7.5",
