+++ src/validations/__test__/acceptance/css/u2p2.css
@@ -1,0 +1,65 @@
+body{
+    background-color: #d9d9d9;
+}
+
+h1{
+    text-align:center;
+    font-size:70px;
+    font-family: cursive;
+    color:#99ffff;
+    
+}
+#name{
+    position: absolute;
+    top:200px;
+    left:300px;
+    color:#ccffff;
+    font-family: cursive;
+    font-size: 30px;
+}
+#job{
+    position: absolute;
+    top:300px;
+    left:300px;
+    font-family: cursive;
+    font-size:30px;
+    color:#ccffff;
+}
+#fun{
+     position: absolute;
+    top:400px;
+    left:300px;
+    font-family: cursive;
+    font-size:30px;
+    color:#ccffff;
+}
+#movie{
+     position: absolute;
+    top:500px;
+    left:300px;
+    font-family: cursive;
+    font-size:30px;
+    color:#ccffff;
+}
+#music{
+     position: absolute;
+    top:600px;
+    left:300px;
+    font-family: cursive;
+    font-size:30px;
+    color:#ccffff;
+}
+#year{
+    position: absolute;
+    top:700px;
+    left:300px;
+    font-family: cursive;
+    font-size:30px;
+    color:#ccffff;
+}
+
+#image{
+    position: absolute;
+    top:400px;
+    left: 1200px;
+}
