+++ src/validations/__test__/acceptance/css/u5p1.css
@@ -1,0 +1,7 @@
+body{
+    margin-bottom: 100px;
+}
+
+div{
+    font-size: 2em;
+}
