+++ src/validations/__test__/acceptance/css/u5p2.css
@@ -1,0 +1,45 @@
+
+#gallery {
+  width: 50%;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.gallery-thumb {
+  display: block;
+  float: left;
+  padding: 5px;
+  box-sizing: border-box;
+  width: 33%;
+}
+
+#change-color {
+  width: 300px;
+  text-align: center;
+  display: block;
+  margin: 0 auto;
+  font-size: 3em;
+}
+
+.section {
+  padding: 1em 0;
+  border-bottom: 2px dashed gray;
+}
+
+h1, #submit-name {
+  font-size: 2em;
+}
+
+input {
+  font-size: 1em;
+}
+
+#greeting {
+  text-align: center;
+  font-size: 2em;
+  text-shadow: 3px 3px 3px gray;
+}
+
+#gallery-main {
+  width: 100%;
+}
