+++ src/validations/__test__/acceptance/html/u1p1.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <title>Unit 1, Project 1</title>
+    </head>
+    <body>
+        <h1>I love ScriptEd!</h1>
+        <img src="http://i.imgur.com/J5ji6zX.png">
+        <p>This is the ScriptEd logo</p>
+        <a href="www.scripted.org">This is a link to ScriptEd's website </a>
+        
+        <img src="http://i.imgur.com/EwUGnTi.png">
+        <p>In ScriptEd we learn the HTML, CSS and JavaScript</p>
+        <a href="https://github.com/ScriptEdcurriculum/curriculum2016">This is a link to ScriptEd's curriculum </a>
+    </body>
+</html>
