+++ src/validations/__test__/acceptance/html/u2p2.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <title>Grace Hopper</title>
+        <link rel="stylesheet" type="text/css" href="style.css">
+    </head>
+    <body>
+        <h1 id="title">This page is about me</h1>
+        <p id="name">My name is Grace Hopper</p>
+        <p id="job">I work in the NAVY as a programmer</p>
+        <p id="fun">For fun, I research how to write new machine programs</p>
+        <p id="movie">My favorite movie is "Hamlet"</p>
+        <p id="music">My favorite type of music is jazz</p>
+        <p id="year">In ten years I hope to be creating more complex programs that help the Navy</p>
+        <a href="https://en.wikipedia.org/wiki/Grace_Hopper">Here is a link to my Wikipedia page</a>
+        <div id="image"><img src="https://media1.britannica.com/eb-media/00/149300-004-D49D0778.jpg"></div>
+    </body>
+</html>
