+++ src/validations/__test__/acceptance/html/u4p1.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <title>ScriptEd CSS Garden</title>
+    </head>
+  <body>
+    <h1 id="banner">ScriptEd CSS Site!</h1>
+    <h1 id="tagline">We love ScriptEd</h1>
+    <img src="http://i65.tinypic.com/qxvr46.jpg">
+
+     <p id="mission">ScriptEd equips students with the fundamental coding skills and professional experiences that together create access to careers in technology.</p>
+      <p id="about"> This year in ScriptEd we are going to create several websites, games and webapps! </p>
+      <h1 id="topics"> In ScriptEd we learn...</h1>
+      <ul>
+        <li>HTML</li>
+        <li>CSS</li>
+        <li>JavaScript</li>
+        <li>Git</li>
+        <li>jQuery</li>
+      </ul>
+  </body>
+</html>
