+++ src/validations/__test__/acceptance/html/u5p1.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <title>jQuery Puns</title>
+    </head>
+    <body>
+    <h1>ScriptEd Pun-A-Thon</h1>
+    <hr>
+    <div id="joke-1"> What kind of cheese doesn't belong to you?
+    <button>Click</button>
+    </div>
+    <hr>
+    <div id="joke-2"> What does the annoying pepper do? 
+    <button>Click</button>
+    </div>
+    <hr>
+        <div id="joke-3">
+        <p> Why did the bike fall over? </p>
+        <button>Click</button>
+    </div>
+    <hr>
+        <img id="joke-4" src="http://i.imgur.com/tWNHJEV.jpg">
+</body>
+</html>
