+++ src/validations/__test__/acceptance/html/u5p2.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <script src="https://code.jquery.com/jquery-2.1.4.js"></script>
+        <script src="script.js"></script>
+        <link rel="stylesheet" type="text/css" href="style.css">
+        <title>jQuery Fun House Solution Code</title>
+    </head>
+    <body>
+        <div class="section">
+      <h1>What is your name?
+        <input id="name">
+      </h1>
+      <div>
+        <button id="submit-name" value="That’s my name">That's my name</button>
+      </div>
+      <p id="greeting">Hi </p>
+    </div>
+
+    <div class="section">
+      <button id="change-color">Change the page's color!</button>
+    </div>
+
+    <div class="section" id="gallery">
+      <h1>Pictures of elephants</h1>
+      <img class="gallery-thumb" id="pic1" src="https://upload.wikimedia.org/wikipedia/commons/f/f9/Loxodonta_africana_-_old_bull_(Ngorongoro,_2009).jpg">
+      <img class="gallery-thumb" id="pic2" src="https://upload.wikimedia.org/wikipedia/commons/6/63/African_elephant_warning_raised_trunk.jpg">
+      <img class="gallery-thumb" id="pic3" src="https://iso.500px.com/wp-content/uploads/2014/08/2048-5.jpg">
+      <img id="gallery-main" src="www.fake.com">
+    </div>
+    </body>
+</html>
