--- test/helpers/testValidatorAcceptance.js
@@ -1,11 +1,0 @@
-import acceptance from '../data/acceptance.json';
-
-import validationTest from './validationTest';
-
-export default function testValidatorAcceptance(validator, language) {
-  return t => {
-    acceptance[language].forEach(source =>
-      t.test(`Acceptance - ${language}`, validationTest(source, validator)),
-    );
-  };
-}
